# backend/docs/swagger.yaml
basePath: /v1
definitions:
  dto.ActionResponse:
    properties:
      id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  dto.CreateReminderRequest:
    properties:
      body:
        type: string
      locale:
        type: string
      recipient:
        type: string
      scheduledAt:
        type: string
      subject:
        type: string
      templateId:
        type: string
      templateKey:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      userId:
        type: string
      variables:
        additionalProperties:
          type: string
        type: object
    required:
    - recipient
    - scheduledAt
    - type
    - userId
    type: object
...
  dto.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
      requestId:
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.ReminderResponse:
    properties:
      body:
        type: string
      cancelledAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      locale:
        type: string
      notificationId:
        type: string
      recipient:
        type: string
      scheduledAt:
        type: string
      status:
        type: string
      subject:
        type: string
      templateId:
        type: string
      templateKey:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      updatedAt:
        type: string
      userId:
        type: string
      variables:
        additionalProperties:
          type: string
        type: object
    type: object
...
  models.NotificationType:
    enum:
    - sms
    - email
    - push
    - in_app
    - webhook
    - security
    type: string
    x-enum-varnames:
    - NotificationTypeSMS
    - NotificationTypeEmail
    - NotificationTypePush
    - NotificationTypeInApp
    - NotificationTypeWebhook
    - NotificationTypeSecurity
...
host: 127.0.0.1:9002
info:
  contact:
    email: support@minisource.com
    name: API Support
  description: Notification Service for Minisource - Handles Email, SMS, Push, and
    In-App notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Notifier Service API
  version: "2.0"
paths:
...
  /reminders:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of reminders with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Start of scheduled date range (ISO8601)
        in: query
        name: scheduledFrom
        type: string
      - description: End of scheduled date range (ISO8601)
        in: query
        name: scheduledTo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List reminders
      tags:
      - Reminders
    post:
      consumes:
      - application/json
      description: Create a new scheduled reminder
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      - description: Reminder data
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create reminder
      tags:
      - Reminders
  /reminders/{reminderId}:
    delete:
      consumes:
      - application/json
      description: Delete a scheduled or cancelled reminder
      parameters:
      - description: Reminder ID
        in: path
        name: reminderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete reminder
      tags:
      - Reminders
    get:
      consumes:
      - application/json
      description: Retrieve a specific reminder
      parameters:
      - description: Reminder ID
        in: path
        name: reminderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get reminder by ID
      tags:
      - Reminders
    put:
      consumes:
      - application/json
      description: Update an existing scheduled reminder
      parameters:
      - description: Reminder ID
        in: path
        name: reminderId
        required: true
        type: string
      - description: Reminder data
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update reminder
      tags:
      - Reminders
  /reminders/{reminderId}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a scheduled or processing reminder
      parameters:
      - description: Reminder ID
        in: path
        name: reminderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel reminder
      tags:
      - Reminders
  /reminders/user/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieve reminders for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user reminders
      tags:
      - Reminders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"